backend/requirements.txt
application-properties==0.6.0
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
Columnar==1.4.1
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-colorfield==0.9.0
django-filter==23.2
django-import-export==3.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.5.0
et-xmlfile==1.1.0
filetype==1.2.0
flake8==6.0.0
flake8-commas==2.1.0
flake8-docstrings==1.7.0
Flake8-pyproject==1.2.3
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
MarkupPy==1.14
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==2.7.0
pymarkdown==0.1.4
pymarkdownlnt==0.9.11
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
urllib3==2.0.3
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
